Add modified and untracked files

+++ ComInt.java
@@ -7,8 +7,8 @@ public class ComInt {
 
         double p;
         double r;
-        double n;
-        double t;
+        int n;
+        int t;
 
         
         System.out.print("Enter Principal: ");
@@ -18,14 +18,14 @@ public class ComInt {
         r = scanner.nextDouble() / 100.00;
 
         System.out.print("Enter how many times it is compounded annaully: ");
-        n = scanner.nextDouble(); 
+        n = scanner.nextInt(); 
         
         System.out.print("Enter for how much time you want to invest: ");
-        t = scanner.nextDouble();
+        t = scanner.nextInt();
 
         double amount = p * Math.pow((1 + (r / n)), n*t);
 
-        System.out.printf("Compound Interest after investing it for %.2f years with rate of interest of %.2f is %.2f ", t, r, amount);
+        System.out.printf("Compound Interest after investing it for %d years with rate of interest of %.2f is %.2f ", t, r, amount);
 
         scanner.close();
     }
